Common code reuse between BLS12-377 and BLS12-388, remove excess code
