+++ libff/algebra/curves/bls12_381/bls12_381_pp.cpp
@@ -1,20 +1,33 @@
+/** @file
+ *****************************************************************************
+ * @author     This file is part of libff, developed by SCIPR Lab
+ *             and contributors (see AUTHORS).
+ * @copyright  MIT license (see LICENSE file)
+ *****************************************************************************/
+
 #include <libff/algebra/curves/bls12_381/bls12_381_pp.hpp>
 #include <libff/algebra/curves/bls12.tcc>
 
+
 namespace libff {
+
 
 bigint<bls12_381_r_limbs> bls12_381_modulus_r;
 bigint<bls12_381_q_limbs> bls12_381_modulus_q;
 
-bls12_381_Fq bls12_381_coeff_b;
-bls12_381_Fq2 bls12_381_twist;
-bls12_381_Fq2 bls12_381_twist_coeff_b;
-/*
-bls12_381_Fq2 bls12_381_twist_mul_by_q_X;
-bls12_381_Fq2 bls12_381_twist_mul_by_q_Y;
-*/
 
-bls12_381_pp::Fqe_type bls12_381_pp::TWIST_COEFF_B;
+bls12_381_G1 bls12_381_G1::_one;
+bls12_381_G1 bls12_381_G1::_zero;
+bls12_381_Fq bls12_381_G1::coeff_b;
+std::vector<size_t> bls12_381_G1::wnaf_window_table;
+std::vector<size_t> bls12_381_G1::fixed_base_exp_window_table;
+
+
+bls12_381_G2 bls12_381_G2::_one;
+bls12_381_G2 bls12_381_G2::_zero;
+bls12_381_Fq2 bls12_381_G2::coeff_b;
+std::vector<size_t> bls12_381_G2::wnaf_window_table;
+std::vector<size_t> bls12_381_G2::fixed_base_exp_window_table;
 
 
 void bls12_381_pp::init_public_params()
@@ -119,9 +132,9 @@ void bls12_381_pp::init_public_params()
     bls12_381_Fq12::Frobenius_coeffs_c1[11] = bls12_381_Fq2(bls12_381_Fq("877076961050607968509681729531255177986764537961432449499635504522207616027455086505066378536590128544573588734230"), bls12_381_Fq("3125332594171059424908108096204648978570118281977575435832422631601824034463382777937621250592425535493320683825557"));
     
     /* choice of short Weierstrass curve and its twist */
-    bls12_381_coeff_b = bls12_381_Fq("4");
-    bls12_381_twist = bls12_381_Fq2::nqr;
-    bls12_381_pp::TWIST_COEFF_B = bls12_381_twist_coeff_b = bls12_381_coeff_b * bls12_381_twist;  // M type twist, Fq2::b should be [4,4]
+    bls12_381_G1::coeff_b = bls12_381_Fq("4");
+    //bls12_381_twist = bls12_381_Fq2::nqr;
+    bls12_381_G2::coeff_b = bls12_381_G1::coeff_b * bls12_381_Fq2::nqr;  // M type twist, Fq2::b should be [4,4]
     //bls12_381_twist_mul_by_b_c0 = bls12_381_coeff_b * bls12_381_Fq2::non_residue;
     //bls12_381_twist_mul_by_b_c1 = bls12_381_coeff_b * bls12_381_Fq2::non_residue;
 
@@ -197,16 +210,20 @@ void bls12_381_pp::init_public_params()
 
 
     /* choice of group G2 */
-    bls12_381_G2::_zero = bls12_381_G2(bls12_381_Fq2::zero(),
-                                         bls12_381_Fq2::one(),
-                                         bls12_381_Fq2::zero());
+    bls12_381_G2::_zero = bls12_381_G2(
+                            bls12_381_Fq2::zero(),
+                            bls12_381_Fq2::one(),
+                            bls12_381_Fq2::zero());
 
     // simple G2 generator
-    bls12_381_G2::_one = bls12_381_G2(bls12_381_Fq2(bls12_381_Fq("352701069587466618187139116011060144890029952792775240219908644239793785735715026873347600343865175952761926303160"),
-                                                      bls12_381_Fq("3059144344244213709971259814753781636986470325476647558659373206291635324768958432433509563104347017837885763365758")),
-                                        bls12_381_Fq2(bls12_381_Fq("1985150602287291935568054521177171638300868978215655730859378665066344726373823718423869104263333984641494340347905"),
-                                                      bls12_381_Fq("927553665492332455747201965776037880757740193453592970025027978793976877002675564980949289727957565575433344219582")),
-                                        bls12_381_Fq2::one());
+    bls12_381_G2::_one = bls12_381_G2(
+                            bls12_381_Fq2(
+                                bls12_381_Fq("352701069587466618187139116011060144890029952792775240219908644239793785735715026873347600343865175952761926303160"),
+                                bls12_381_Fq("3059144344244213709971259814753781636986470325476647558659373206291635324768958432433509563104347017837885763365758")),
+                            bls12_381_Fq2(
+                                bls12_381_Fq("1985150602287291935568054521177171638300868978215655730859378665066344726373823718423869104263333984641494340347905"),
+                                bls12_381_Fq("927553665492332455747201965776037880757740193453592970025027978793976877002675564980949289727957565575433344219582")),
+                            bls12_381_Fq2::one());
 
 
     // // TODO: wNAF window table
@@ -263,20 +280,11 @@ void bls12_381_pp::init_public_params()
     // // window 22 is unbeaten in [31673814.95, inf]
     // bls12_381_G2::fixed_base_exp_window_table.push_back(31673815);
 
-
-
-    /* pairing parameters */
-    // TODO: final_exp_z=last_chunk? 
     /* 
      * https://eprint.iacr.org/2017/1174.pdf
      * ate_loop_count=t where t in the value chosen in q(t) and r(t) parameterization of BLS curve.
      * for BLS12_377 t=3·2^46·(7·13·499)+1 s.t. t=1 (mod 3·2^46) to have a high 2-adicity for q and r. 
     */
-    //bls12_381_ate_loop_count = bigint_q("15132376222941642752");
-    //bls12_381_ate_is_loop_count_neg = true; 
-    //bls12_381_final_exponent = bigint<12*bls12_381_q_limbs>("322277361516934140462891564586510139908379969514828494218366688025288661041104682794998680497580008899973249814104447692778988208376779573819485263026159588510513834876303014016798809919343532899164848730280942609956670917565618115867287399623286813270357901731510188149934363360381614501334086825442271920079363289954510565375378443704372994881406797882676971082200626541916413184642520269678897559532260949334760604962086348898118982248842634379637598665468817769075878555493752214492790122785850202957575200176084204422751485957336465472324810982833638490904279282696134323072515220044451592646885410572234451732790590013479358343841220074174848221722017083597872017638514103174122784843925578370430843522959600095676285723737049438346544753168912974976791528535276317256904336520179281145394686565050419250614107803233314658825463117900250701199181529205942363159325765991819433914303908860460720581408201373164047773794825411011922305820065611121544561808414055302212057471395719432072209245600258134364584636810093520285711072578721435517884103526483832733289802426157301542744476740008494780363354305116978805620671467071400711358839553375340724899735460480144599782014906586543813292157922220645089192130209334926661588737007768565838519456601560804957985667880395221049249803753582637708560");
-    //bls12_381_final_exponent_z = bigint_q("15132376222941642752");
-    //bls12_381_final_exponent_is_z_neg = true;
 }
