+++ libff/algebra/curves/bls12_381/bls12_381_pp.hpp
@@ -9,16 +9,17 @@
 #define BLS12_381_PP_HPP_
 
 #include <libff/algebra/curves/bls12.hpp>
+#include <libff/algebra/curves/sw.hpp>
 #include <libff/algebra/fields/fp12_2over3over2.hpp>
 
 namespace libff {
 
 
-const mp_size_t bls12_381_r_bitcount = 255;
-const mp_size_t bls12_381_q_bitcount = 381;
+constexpr mp_size_t bls12_381_r_bitcount = 255;
+constexpr mp_size_t bls12_381_q_bitcount = 381;
 
-const mp_size_t bls12_381_r_limbs = (bls12_381_r_bitcount+GMP_NUMB_BITS-1)/GMP_NUMB_BITS;
-const mp_size_t bls12_381_q_limbs = (bls12_381_q_bitcount+GMP_NUMB_BITS-1)/GMP_NUMB_BITS;
+constexpr mp_size_t bls12_381_r_limbs = (bls12_381_r_bitcount+GMP_NUMB_BITS-1)/GMP_NUMB_BITS;
+constexpr mp_size_t bls12_381_q_limbs = (bls12_381_q_bitcount+GMP_NUMB_BITS-1)/GMP_NUMB_BITS;
 
 extern bigint<bls12_381_r_limbs> bls12_381_modulus_r;
 extern bigint<bls12_381_q_limbs> bls12_381_modulus_q;
@@ -30,18 +31,10 @@ typedef Fp6_3over2_model<bls12_381_q_limbs, bls12_381_modulus_q> bls12_381_Fq6;
 typedef Fp12_2over3over2_model<bls12_381_q_limbs, bls12_381_modulus_q> bls12_381_Fq12;
 typedef bls12_381_Fq12 bls12_381_GT;
 
-// parameters for the curve E/Fq : y^2 = x^3 + b
-extern bls12_381_Fq bls12_381_coeff_b;
-// parameters for the twisted curve E'/Fq2 : y^2 = x^3 + b/xi
-extern bls12_381_Fq2 bls12_381_twist;
-extern bls12_381_Fq2 bls12_381_twist_coeff_b;
-extern bls12_381_Fq bls12_381_twist_mul_by_b_c0;
-extern bls12_381_Fq bls12_381_twist_mul_by_b_c1;
-extern bls12_381_Fq2 bls12_381_twist_mul_by_q_X;
-extern bls12_381_Fq2 bls12_381_twist_mul_by_q_Y;
 
 class bls12_381_G1;
 class bls12_381_G2;
+
 
 class bls12_381_pp {
 public:
@@ -61,8 +54,6 @@ public:
     static constexpr auto X_NUM_ONES = bls12::CountOnes<X>::n;
     static constexpr bool X_IS_NEG = true;
 
-    static Fqe_type TWIST_COEFF_B;
-
     static void init_public_params();
     static G1_precomp_type precompute_G1(const G1_type &P);
     static G2_precomp_type precompute_G2(const G2_type &Q);
@@ -77,9 +68,55 @@ public:
     static Fqk_type reduced_pairing(const G1_type &P, const G2_type &Q);
 };
 
+
+using bls12_381_G1_base = sw::SWJacobianPoint<bls12_381_pp, bls12_381_pp::Fq_type, bls12_381_pp::Fp_type, bls12_381_G1>;
+
+class bls12_381_G1 : public bls12_381_G1_base
+{
+    using bls12_381_G1_base::SWJacobianPoint;
+
+public:
+    // Pass-thru Copy constructor from base type
+    bls12_381_G1( const bls12_381_G1_base &base )
+    : bls12_381_G1_base(base) {}
+
+    int sign_bit() const {
+        return Y.as_bigint().data[0] & 1;
+    }
+
+    // Define storage in our own class, otherwise it gets messy
+    static base_field coeff_b;
+    static std::vector<size_t> wnaf_window_table;
+    static std::vector<size_t> fixed_base_exp_window_table;
+    static bls12_381_G1 _zero;
+    static bls12_381_G1 _one;
+};
+
+
+using bls12_381_G2_base = sw::SWJacobianPoint<bls12_381_pp, bls12_381_pp::Fqe_type, bls12_381_pp::Fp_type, bls12_381_G2>;
+
+class bls12_381_G2 : public bls12_381_G2_base
+{
+    using bls12_381_G2_base::SWJacobianPoint;
+
+public:
+    // Pass-thru Copy constructor from base type
+    bls12_381_G2( const bls12_381_G2_base &base )
+    : bls12_381_G2_base(base) {}
+
+    int sign_bit() const {
+        return Y.c0.as_bigint().data[0] & 1;
+    }
+
+    // Define storage in our own class, otherwise it gets messy
+    static base_field coeff_b;
+    static std::vector<size_t> wnaf_window_table;
+    static std::vector<size_t> fixed_base_exp_window_table;
+    static bls12_381_G2 _zero;
+    static bls12_381_G2 _one;
+};
+
+
 } // libff
 
 #endif // BLS12_381_PP_HPP_
-
-#include <libff/algebra/curves/bls12_381/bls12_381_g1.hpp>
-#include <libff/algebra/curves/bls12_381/bls12_381_g2.hpp>
